fix(network): use verified https by default
SECURITY.md was wrong/outdated on stating that we use http. The code
enforces https. The issue was that the certificate was not being
verified. Now, by default, it is.

Still, you can skip it with a configuration variable for development
purposes when you use self signed certificates.

fixes https://github.com/SUSE/spacewalk/issues/27831

Signed-off-by: Jordi Massaguer Pla <jmassaguerpla@suse.com>